--- CJS/Assets/Scripts/CoinTile.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d78e296656758480b86f958db568947a
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
