--- CJS/Assets/Scripts/GroundTile.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 553bf7947dfe44fae8eaf9beaaeb14c1
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
