--- CJS/Assets/Scripts/House.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 3593db181cd64473f8ccde981b825583
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
