--- CJS/Assets/Scripts/HouseTile.cs
@@ -1,7 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-using UnityEngine.Tilemaps;
-
-public class HouseTile: Tile {
-}
