--- CJS/Assets/Scripts/HouseTile.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 8762ce50457a848c0a3f0989d46699a9
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
