+++ CJS/Assets/Scripts/Poi.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: b65c9ad1b3718427ebc0dd94e9ef7069
+guid: fb3f10396f427439f8d41cf1ee9f8d44
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
