+++ CJS/ProjectSettings/ProjectVersion.txt
@@ -1,2 +1,2 @@
-m_EditorVersion: 2019.4.9f1
-m_EditorVersionWithRevision: 2019.4.9f1 (50fe8a171dd9)
+m_EditorVersion: 2019.4.10f1
+m_EditorVersionWithRevision: 2019.4.10f1 (5311b3af6f69)
